+++ CHANGELOG.md
@@ -13,7 +13,13 @@
 * Increment vCD API version used from 27.0 to 29.0
     * Remove fields `VdcEnabled`, `VAppParentHREF`, `VAppParentName`, `HighestSupportedVersion`, `VmToolsVersion`, `TaskHREF`, `TaskStatusName`, `TaskDetails`, `TaskStatus` from `QueryResultVMRecordType`
     * Add fields `ID, Type, ContainerName, ContainerID, OwnerName, Owner, NetworkHref, IpAddress, CatalogName, VmToolsStatus, GcStatus, AutoUndeployDate, AutoDeleteDate, AutoUndeployNotified, AutoDeleteNotified, Link, MetaData` to `QueryResultVMRecordType`, `DistributedInterface` to `NetworkConfiguration` and `RegenerateBiosUuid` to `VMGeneralParams`
-    * Change to pointers `DistributedRoutingEnabled` in `GatewayConfiguration` and `DistributedInterface` in `NetworkConfiguration`
+    * Change to pointers `DistributedRoutingEnabled` in `GatewayConfiguration` and
+    `DistributedInterface` in `NetworkConfiguration`
+* Add new field to type `GatewayConfiguration`: `FipsModeEnabled` -
+  [#267](https://github.com/vmware/go-vcloud-director/pull/267)
+* Change bool to bool pointer for fields in type `GatewayConfiguration`: `HaEnabled`,
+  `UseDefaultRouteForDNSRelay`, `AdvancedNetworkingEnabled` -
+  [#267](https://github.com/vmware/go-vcloud-director/pull/267)
 * Added method `EdgeGateway.GetLbVirtualServers` that gets all virtual servers configured on NSX load balancer. [#266](https://github.com/vmware/go-vcloud-director/pull/266)
 * Added method `EdgeGateway.GetLbServerPools` that gets all pools configured on NSX load balancer. [#266](https://github.com/vmware/go-vcloud-director/pull/266)
 * Added method `EdgeGateway.GetLbServiceMonitors` that gets all service monitors configured on NSX load balancer. [#266](https://github.com/vmware/go-vcloud-director/pull/266)